fix(object_merger): move headers to object_merger directory
