+++ perception/object_merger/include/object_merger/data_association/solver/gnn_solver.hpp
@@ -15,8 +15,8 @@
 #ifndef OBJECT_ASSOCIATION_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
 #define OBJECT_ASSOCIATION_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
 
-#include "object_association_merger/data_association/solver/gnn_solver_interface.hpp"
-#include "object_association_merger/data_association/solver/mu_successive_shortest_path.hpp"
-#include "object_association_merger/data_association/solver/successive_shortest_path.hpp"
+#include "object_merger/data_association/solver/gnn_solver_interface.hpp"
+#include "object_merger/data_association/solver/mu_successive_shortest_path.hpp"
+#include "object_merger/data_association/solver/successive_shortest_path.hpp"
 
 #endif  // OBJECT_ASSOCIATION_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
