+++ perception/object_merger/src/object_association_merger/data_association/mu_successive_shortest_path/mu_successive_shortest_path_wrapper.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "object_association_merger/data_association/solver/mu_successive_shortest_path.hpp"
+#include "object_merger/data_association/solver/mu_successive_shortest_path.hpp"
 
 #include <mussp/mussp.h>
